+++ src/onemail/om-common/src/types/index.ts
@@ -1,1 +1,1 @@
-export * from './EmailStatusHistory.js';
+export * from './emailStatusHistory.js';